moved the read the docs based codebase summaries to codebase.py and codebase_utils.py
